requirements.txt
django==1.4.5
South
psycopg2
django-flickr
